+++ src/main/java/cc/cassian/rrv/common/builtin/BuiltInReliableRecipeViewerIntegration.java
@@ -1,19 +1,16 @@
 package cc.cassian.rrv.common.builtin;
 
 import cc.cassian.rrv.api.CommonTags;
-import cc.cassian.rrv.common.builtin.tag.TagClientRecipe;
+import cc.cassian.rrv.common.builtin.info.InfoServerRecipe;
 import cc.cassian.rrv.common.builtin.tag.TagServerRecipe;
 import com.mojang.datafixers.util.Either;
 import cc.cassian.rrv.api.ReliableRecipeViewerPlugin;
 import cc.cassian.rrv.api.recipe.ItemView;
 import cc.cassian.rrv.common.builtin.blasting.BlastingServerRecipe;
 import cc.cassian.rrv.common.builtin.brewing.BrewingServerRecipe;
-import cc.cassian.rrv.common.builtin.brewing.BrewingClientRecipe;
 import cc.cassian.rrv.common.builtin.burning.BurningServerRecipe;
-import cc.cassian.rrv.common.builtin.burning.BurningClientRecipe;
 import cc.cassian.rrv.common.builtin.campfire.CampfireServerRecipe;
 import cc.cassian.rrv.common.builtin.entity.EntityServerRecipe;
-import cc.cassian.rrv.common.builtin.entity.EntityClientRecipe;
 import cc.cassian.rrv.common.builtin.shaped.ShapedServerRecipe;
 import cc.cassian.rrv.common.builtin.shapeless.ShapelessServerRecipe;
 import cc.cassian.rrv.common.builtin.smelting.SmeltingServerRecipe;
@@ -22,7 +19,6 @@ import cc.cassian.rrv.common.builtin.smoking.SmokingServerRecipe;
 import cc.cassian.rrv.common.builtin.stonecutting.StonecutterServerRecipe;
 import cc.cassian.rrv.common.builtin.tipped_arrow.TippedArrowServerRecipe;
 import cc.cassian.rrv.common.builtin.transmute.TransmuteServerRecipe;
-import cc.cassian.rrv.common.builtin.villager.VillagerClientRecipe;
 import cc.cassian.rrv.common.builtin.villager.VillagerServerRecipe;
 import cc.cassian.rrv.common.mixin.world.item.alchemy.PotionBrewingAccessor;
 import cc.cassian.rrv.common.mixin.world.item.crafting.IngredientAccessor;
@@ -35,14 +31,6 @@ import cc.cassian.rrv.common.mixin.world.level.storage.loot.entries.LootPoolSing
 import cc.cassian.rrv.common.mixin.world.level.storage.loot.functions.SetPotionFunctionAccessor;
 import cc.cassian.rrv.common.recipe.ServerRecipeManager;
 import cc.cassian.rrv.common.recipe.inventory.SlotContent;
-import cc.cassian.rrv.common.builtin.blasting.BlastingClientRecipe;
-import cc.cassian.rrv.common.builtin.campfire.CampfireClientRecipe;
-import cc.cassian.rrv.common.builtin.shaped.CraftingClientRecipe;
-import cc.cassian.rrv.common.builtin.shapeless.ShapelessClientRecipe;
-import cc.cassian.rrv.common.builtin.smelting.SmeltingClientRecipe;
-import cc.cassian.rrv.common.builtin.smithing.SmithingClientRecipe;
-import cc.cassian.rrv.common.builtin.smoking.SmokingClientRecipe;
-import cc.cassian.rrv.common.builtin.stonecutting.StonecutterClientRecipe;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.Holder;
 import net.minecraft.core.HolderSet;
@@ -71,9 +59,7 @@ import net.minecraft.world.item.alchemy.PotionContents;
 import net.minecraft.world.item.component.ItemLore;
 import net.minecraft.world.item.crafting.*;
 import net.minecraft.world.item.enchantment.*;
-import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.entity.FuelValues;
-import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.entries.CompositeEntryBase;
@@ -101,15 +87,6 @@ public class BuiltInReliableRecipeViewerIntegration implements ReliableRecipeVie
 
     @Override
     public void onIntegrationInitialize() {
-
-
-        ItemView.addClientReloadCallback(() -> {
-
-            BuiltInRegistries.BLOCK.get(CommonTags.EXCLUDED_BLOCKS).ifPresent(blocks -> blocks.stream().filter(Holder::isBound).filter(Holder::isBound).map(Holder::value).forEach(block -> ItemView.excludeItem(block.asItem())));
-            BuiltInRegistries.ITEM.get(CommonTags.EXCLUDED_ITEMS).ifPresent(items -> items.stream().filter(Holder::isBound).filter(Holder::isBound).map(Holder::value).forEach(ItemView::excludeItem));
-            BuiltInRegistries.FLUID.get(CommonTags.EXCLUDED_FLUIDS).ifPresent(fluids -> fluids.stream().filter(Holder::isBound).filter(Holder::isBound).map(Holder::value).forEach(fluid -> ItemView.excludeItem(fluid.defaultFluidState().createLegacyBlock().getBlock().asItem())));
-
-        });
 
         ItemView.addServerReloadCallback(() -> {
 
@@ -149,7 +126,9 @@ public class BuiltInReliableRecipeViewerIntegration implements ReliableRecipeVie
 
         //providers
 
+
         ItemView.addServerRecipeProvider(recipeList -> {
+            recipeList.add(new InfoServerRecipe());
 
             BuiltInRegistries.ENTITY_TYPE.forEach(entityType -> {
                 if (entityType.getDefaultLootTable().isEmpty())
@@ -204,8 +183,8 @@ public class BuiltInReliableRecipeViewerIntegration implements ReliableRecipeVie
             });
 
             BuiltInRegistries.ITEM.listTagIds().forEach((tag) -> {
-				Optional<HolderSet.Named<Item>> tagContents = BuiltInRegistries.ITEM.get(tag);
-                if (tagContents.isPresent() && !tagContents.get().stream().allMatch(item-> ItemView.getExcludedItems().contains(item.value()))) {
+                Optional<HolderSet.Named<Item>> tagContents = BuiltInRegistries.ITEM.get(tag);
+                if (tagContents.isPresent() && !tagContents.get().stream().allMatch(item -> ItemView.getExcludedItems().contains(item.value()))) {
                     recipeList.add(new TagServerRecipe(tag));
                 }
             });
@@ -427,53 +406,6 @@ public class BuiltInReliableRecipeViewerIntegration implements ReliableRecipeVie
 
         });
         *///?}
-
-        //Wrapper
-        ItemView.addClientRecipeWrapper(BurningServerRecipe.TYPE, unwrapped -> List.of(new BurningClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(SmeltingServerRecipe.TYPE, unwrapped -> List.of(new SmeltingClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(BlastingServerRecipe.TYPE, unwrapped -> List.of(new BlastingClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(SmokingServerRecipe.TYPE, unwrapped -> List.of(new SmokingClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(ShapelessServerRecipe.TYPE, unwrapped -> List.of(new ShapelessClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(ShapedServerRecipe.TYPE, unwrapped -> List.of(new CraftingClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(TransmuteServerRecipe.TYPE, unwrapped -> List.of(new ShapelessClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(TippedArrowServerRecipe.TYPE, unwrapped -> List.of(new CraftingClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(CampfireServerRecipe.TYPE, unwrapped -> List.of(new CampfireClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(StonecutterServerRecipe.TYPE, unwrapped -> List.of(new StonecutterClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(SmithingServerRecipe.TYPE, unwrapped -> {
-            List<SmithingClientRecipe> recipes = new ArrayList<>();
-
-            if (unwrapped.getBase() != null && unwrapped.getTemplate() != null) {
-                SlotContent.of(unwrapped.getBase()).getValidContents().forEach(baseStack -> {
-                    SlotContent.of(unwrapped.getTemplate()).getValidContents().forEach(templateStack -> {
-                        recipes.add(new SmithingClientRecipe(unwrapped.isTrim(), unwrapped.getAddition(), Ingredient.of(baseStack.getItem()), Ingredient.of(templateStack.getItem()), unwrapped.getPattern(), unwrapped.getUpgradeResult()));
-                    });
-
-                });
-            } else if (unwrapped.getBase() != null && unwrapped.getAddition() != null) {
-                SlotContent.of(unwrapped.getBase()).getValidContents().forEach(baseStack -> {
-                    SlotContent.of(unwrapped.getAddition()).getValidContents().forEach(additionStack -> {
-                        recipes.add(new SmithingClientRecipe(unwrapped.isTrim(), Ingredient.of(additionStack.getItem()), Ingredient.of(baseStack.getItem()), unwrapped.getTemplate(), unwrapped.getPattern(), unwrapped.getUpgradeResult()));
-                    });
-
-                });
-            } else if (unwrapped.getAddition() != null && unwrapped.getTemplate() != null) {
-                SlotContent.of(unwrapped.getTemplate()).getValidContents().forEach(templateStack -> {
-                    SlotContent.of(unwrapped.getAddition()).getValidContents().forEach(additionStack -> {
-                        recipes.add(new SmithingClientRecipe(unwrapped.isTrim(), Ingredient.of(additionStack.getItem()), unwrapped.getBase(), Ingredient.of(templateStack.getItem()), unwrapped.getPattern(), unwrapped.getUpgradeResult()));
-                    });
-
-                });
-            }
-
-
-            return recipes;
-        });
-        ItemView.addClientRecipeWrapper(BrewingServerRecipe.TYPE, unwrapped -> List.of(new BrewingClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(VillagerServerRecipe.TYPE, unwrapped -> {
-            return unwrapped.getOffers().stream().map(VillagerClientRecipe::new).toList();
-        });
-        ItemView.addClientRecipeWrapper(EntityServerRecipe.TYPE, unwrapped -> List.of(new EntityClientRecipe(unwrapped)));
-        ItemView.addClientRecipeWrapper(TagServerRecipe.TYPE, unwrapped -> List.of(new TagClientRecipe(unwrapped)));
     }